# update leaves individually in delete_intersecting