--- a/rust/rope/Cargo.toml
+++ b/rust/rope/Cargo.toml
@@ -17,6 +17,8 @@ regex = "1.0"
 [dev-dependencies]
 serde_test = "^1.0"
 serde_json = "1.0"
+rand = "0.6"
+rand_pcg = "0.1"
 
 [features]
 default = []