## Summary
Instead of creating a completely new Spans in delete_intersecting, only update the parts that need to be updated.

## Related Issues
closes #1167 

## Review Checklist
- [ ] I have responded to reviews and made changes where appropriate.
- [X] I have tested the code with `cargo test --all` / `./rust/run_all_checks`.
- [X] I have updated comments / documentation related to the changes I made.
- [X] I have rebased my PR branch onto xi-editor/master.