@cmyr 2019-04-08
@ahouts Cool, thanks for this! If possible, could you give a quick overview of your thinking & approach here? Will make it easier to do a reasonable review if I have a better high-level sense of how you're considering the problem.

It would also be cool to have some concrete benchmarks available to compare these two implementations?