@ahouts 2019-04-08
Sure thing, while I was implementing the logic of walking the tree and modifying leaves I noticed that you can only really do one of three things to any given leaf: change it, remove it, or make no change. I also noticed that the logic of building a new tree based on these modifications was not dependent on what you are doing on each leaf.
As a result, I made the tree walking and construction a function of Node. It accepts a closure as an argument that performs whatever action to every leaf, returning what should be done with each leaf in the new tree.
Then in order to implement span removal, we just have to create a new leaf with the spans that we did not remove. Edge cases are we removed no spans so we return NoChange, or we removed every span so we return Remove.
One nice thing about this approach is it should only make changes if the mutate_leaf function makes changes, and only the dependent nodes will be changed. It will also flatten the tree as Nodes are removed.