@ahouts 2019-04-08
I can't do benchmarking right now, but maybe I'll have time later today.