@ahouts 2019-04-09
I had to make quite a few changes, as I didn't know there is information associated with the position of a Node in the tree. For example, the actual interval of a leaf is given by summing all previous leaves lengths together.
As a quick fix, I removed the `Remove` variant of `MutateResult` so the structure of the tree (and all implied information) will be preserved.
I also had to do a little refactoring so the closure given to `mutate_leaves` also takes the offset as an argument.

After all of that though, the benchmarks look promising. The "range" in the function name refers to the magnitude of the interval that is being deleted.
Original implementation
```
test test_delete_intersecting_10_000_huge_range ... bench:      23,764 ns/iter (+/- 175)
test test_delete_intersecting_10_000_med_range  ... bench:     149,115 ns/iter (+/- 17,102)
test test_delete_intersecting_10_000_tiny_range ... bench:     164,632 ns/iter (+/- 5,320)
```

New Implementation
```
test test_delete_intersecting_10_000_huge_range ... bench:      27,607 ns/iter (+/- 18,746)
test test_delete_intersecting_10_000_med_range  ... bench:      14,611 ns/iter (+/- 136)
test test_delete_intersecting_10_000_tiny_range ... bench:      18,331 ns/iter (+/- 91)
```