@ahouts 2019-04-11
> I worry that this implementation is a bit _fancy_. That is, it's adding a bunch of code to `tree.rs` that I'm not sure is going to be used again, and some of which even if we do decide to pursue this pattern we might want to skip.
Fair enough, I had assumed that rebuilding all nodes except leaves every modification would be too expensive. In retrospect, there are relatively few such nodes compared to the number of leaves.
I also wasn't certain about adding a new method to Tree, but it does make sense to add that functionality there.

I'll take a look into it, that code certainly looks much simpler.