@ahouts 2019-04-11
Benchmarks on this version are quite a bit better, so it looks like you had the right idea.
```
test test_delete_intersecting_100_huge_range       ... bench:         441 ns/iter (+/- 573)
test test_delete_intersecting_100_med_range        ... bench:         739 ns/iter (+/- 3)
test test_delete_intersecting_100_tiny_range       ... bench:         660 ns/iter (+/- 1)
test test_delete_intersecting_10_000_huge_range    ... bench:      28,893 ns/iter (+/- 12,396)
test test_delete_intersecting_10_000_med_range     ... bench:       6,387 ns/iter (+/- 103)
test test_delete_intersecting_10_000_tiny_range    ... bench:       9,756 ns/iter (+/- 35)
test test_delete_intersecting_1_000_000_huge_range ... bench:   3,629,148 ns/iter (+/- 28,344)
test test_delete_intersecting_1_000_000_med_range  ... bench:     436,883 ns/iter (+/- 8,711)
test test_delete_intersecting_1_000_000_tiny_range ... bench:   1,985,280 ns/iter (+/- 30,374)
```

`LeafIterator` returns `(Node, &Leaf)` because the `Node.get_leaf` method is private. If it was not private, `LeafIterator` could return just a `Node`, and be a real iterator.