@ahouts 2019-04-11
One thing to note, all tests (including one with 10,000 spans) pass, but the test with 1,000,000 spans is off by 2. I don't have time to track it down right now though.