@ahouts 2019-04-13
@cmyr I made the changes you suggested. I had to run `cargo update` to get `rand` to compile.