@cmyr 2019-04-13
Okay cool. I've opened a PR that does a general dependency update, in #1174; wanted to figure out why we had three different versions of `rand` in the graph. Hopefully that will be cleaned up now, and we can rebase this onto that when it lands, which will be whenever it passes CI.